A game called "Foxes and Hens"
There are two strategies in the code. One is "take5" strategy, which is a strategy that waits until there are 5 hens in yard, then gathers. My strategy include the probability of Foxes in the card. I determine to gather the card if the probability of Foxes is more than 0.34 or yard is equal to 4. Also, if there are no Foxes in the card, I choose to wait until the end.