A function to apply Anagrams
An anagram is a rearrangement of the letters in a word to form one or more new words. Function anagrams(), which takes as input a phrase and an optional argument, shortest, which is an integer that specifies the shortest acceptable word. Your function should return a set of all the possible combinations of anagrams.